requirements.txt
discord.py>=2.0.0
python-dotenv
fastapi
uvicorn[standard]
sqlalchemy
requests
psycopg2
httpx